Refactored all entities to use Integer IDs instead of Long. Implemented RecipeResource as first web service -  initial service is GET for all recipe entities.

+++ entity/src/main/java/boozelogger/entity/DistillationLog.java
@@ -29,7 +29,7 @@ public class DistillationLog extends Log {
         this(null, null, null, null, new ArrayList<DistillationLogEntry>());
     }
 
-    public DistillationLog(Long id, String notes, Vessel vessel, Date createdAt, List<DistillationLogEntry> entries) {
+    public DistillationLog(Integer id, String notes, Vessel vessel, Date createdAt, List<DistillationLogEntry> entries) {
         super(id, notes, vessel, createdAt);
         this.entries = entries;
     }
