+++ entity/src/main/java/boozelogger/entity/LogEntry.java
@@ -23,7 +23,7 @@ import java.util.Date;
 })
 public abstract class LogEntry {
 
-    private Long id;
+    private Integer id;
     private Double temperature;
     private String notes;
     private Date createdAt;
@@ -32,7 +32,7 @@ public abstract class LogEntry {
         this(null, null, null, null);
     }
 
-    protected LogEntry(Long id, Double temperature, String notes, Date createdAt) {
+    protected LogEntry(Integer id, Double temperature, String notes, Date createdAt) {
         this.id = id;
         this.temperature = temperature;
         this.notes = notes;
@@ -42,11 +42,11 @@ public abstract class LogEntry {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @JsonProperty
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
